Instead of AsVar in IPASIR use a VarAs scheme
This change includes the definition of polymorphic pattern synonyms.
Unfortunately this necessitates moving all Vars into a single module.
That is because pattern and types in separate modules are not recognized
as child and parent, so the pattern cannot be bundled with the type.
Further (orphan) COMPLETE pragmas where every pattern is in another
module are disallowed. Thus -fno-warn-complete-pattern would further
be necessary if patterns were exported standalone without their
parent types.